Prevent negative epochs by using `u32`

+++ src/common.rs
@@ -18,8 +18,8 @@ use serde_derive::{Deserialize, Serialize};
 // Reject unknown REST JSON params with descriptive message.
 #[serde(deny_unknown_fields)]
 pub struct ReservationRequest {
-    pub start_time: i64,
-    pub end_time: i64,
+    pub start_time: u32,
+    pub end_time: u32,
     pub capacity_amount: u32,
     pub user_id: u32,
 }
@@ -33,7 +33,7 @@ impl ReservationRequest {
     /// - `end_time`: Reservation end time, represented by Unix epoch format.
     /// - `capacity_amount`: Amount of resource the user would like to have allocated.
     /// - `user_id`: Your unique identifier.
-    pub fn new(start_time: i64, end_time: i64, capacity_amount: u32, user_id: u32) -> Self {
+    pub fn new(start_time: u32, end_time: u32, capacity_amount: u32, user_id: u32) -> Self {
         Self {
             start_time,
             end_time,
